Added support for wp_site_icon (since WP 4.3)
Added support for the new WP Site Icon (added since WP 4.3) in the favicon selections.

When a user has added a custom site icon this icon will override the stylesheet icons and the hook in get_favicon_path.

Why also the hook? My thought was that since user actually used this it's safe to say he/she want's to use this icon instead of the theme icon. Feedback welcome!