WP 4.3 site icon update
The functions "has_site_icon" and "get_site_icon_url" aren't compatible with multisites.
Icon only displayed when on that blog, in network or other blog the WP logo showed.

This is a fix for this. (hopefully temporarely)

Note the "FALSE !=" , do not change this to "FALSE !==" since the return value doesn't necessarily will be a false when there is no icon, I think it could be an empty string.